feat(ios): valider le code PIN dès le dernier chiffre saisi
Le bouton "Valider" imposait un geste de plus alors que la longueur du code
suffit à savoir que la saisie est finie. Les quatre écrans PIN se soumettent
désormais seuls, et la touche disparaît du pavé.

Trois sensations distinctes portent l'état : le tap doux de chaque chiffre,
un impact ferme quand le premier code est noté et qu'il faut le retaper, la
notification d'erreur ou de succès à la fin. Une pause laisse le dernier point
s'afficher avant que l'écran ne change — c'est le battement que le bouton
fournissait jusqu'ici.

PinRecoveryViewModel sort de sa vue : co-localisé, il poussait le fichier
au-delà du plafond de 500 lignes de SwiftLint.
